feat: use JSON & Jinja for templating
This is a big change. Now, the data of the CV is all stored on a JSON
file. There is no need for editing the `cv.tex` file -- all changes
_must_ be done on the JSON file, and then `make` is used to:
	- Generate the `.tex` file.
	- Compile the CV (as a PDF).

After a successful build, the pipeline is going to push the file onto
the website, and can be accessed on https://bruno.saboia.it/cv